Add horizontal orientation for cartesian charts
This adds a 'horizontal' property to AreaChart, BarChart,
CandlestickChart, LineChart, and ScatterChart via
CartesianChartPropsMixin, with default text rotation adjusted to be
horizontal.

This is a common need for bar charts, and also not uncommon for scatter
plots and candlestick charts.  While not terribly sensible for line and
area charts, there are at least a few mentions online of people wanting
to do this, and regardless, it seems reasonable to

The chart rotation is achieved using an svg transform on the root
element, while the rest of the code remains mostly unchanged, except
that the horizontal option is passed in props to components that decide
how to rotate text labels based on it (AxisTicks and Label)

A new option for 'yOrient' is added as the default: 'default', which
causes the y-axis position to be left if horizontal=false (as before)
and "right" (visually bottom) if horizontal=true.

Margins specified by the user are treated as the visual margins, and
rotated 90 degrees before being used for SVG calculations.